+++ server/utilities/fsUtility.js
@@ -1,0 +1,17 @@
+import fs from 'fs';
+
+function unlink(path) {
+  return new Promise((resolve, reject) => {
+    fs.unlink(path, (error) => {
+      if (error) {
+        reject(error);
+      }
+
+      resolve();
+    });
+  });
+}
+
+export default {
+  unlink
+};