+++ utilities/testSetup.js
@@ -1,10 +1,10 @@
 /* eslint-disable no-process-env */
 
+require('babel-register')({});
+require('react-tap-event-plugin')();
 require('css-modules-require-hook')({
   generateScopedName: '[local]'
 });
-
-require('babel-register')();
 
 // Setting NODE_ENV to test instead of production because setting it to production will suppress error messaging
 // and propType validation warnings.